feat(new-ui): squash repeated steps to groups #4
