+++ lib/adapters/test-result/testplane/index.ts
@@ -1,17 +1,7 @@
-import path from 'path';
-import _ from 'lodash';
+import {ERROR, FAIL, SUCCESS, TestStatus, UNKNOWN_SESSION_ID, UPDATED} from '../../../constants';
+import {ReporterTestResult} from '../index';
 import type Testplane from 'testplane';
-import type {Test as TestplaneTest} from 'testplane';
-import {ValueOf} from 'type-fest';
-
-import {ERROR, FAIL, SUCCESS, TestStatus, UNKNOWN_SESSION_ID, UPDATED} from '../../constants';
-import {
-    getError,
-    hasUnrelatedToScreenshotsErrors,
-    isImageDiffError,
-    isInvalidRefImageError,
-    isNoRefImageError
-} from '../../common-utils';
+import type {Test as TestplaneTest, Config} from 'testplane';
 import {
     Attachment,
     ErrorDetails,
@@ -25,14 +15,23 @@ import {
     ImageInfoSuccess,
     ImageInfoUpdated,
     TestError,
-    TestplaneSuite,
     TestplaneTestResult,
-    TestStepCompressed,
-    TestStepKey
-} from '../../types';
-import {ReporterTestResult} from './index';
-import {getSuitePath} from '../../plugin-utils';
-import {extractErrorDetails} from './utils';
+    TestStepCompressed
+} from '../../../types';
+import _ from 'lodash';
+import {
+    getError,
+    hasUnrelatedToScreenshotsErrors,
+    isImageDiffError,
+    isInvalidRefImageError,
+    isNoRefImageError
+} from '../../../common-utils';
+import {getSuitePath} from '../../../plugin-utils';
+import {extractErrorDetails} from '../utils';
+import path from 'path';
+import {ValueOf} from 'type-fest';
+
+import {getHistory, wrapSkipComment, getSkipComment} from './history';
 
 export const getStatus = (eventName: ValueOf<Testplane['events']>, events: Testplane['events'], testResult: TestplaneTestResult): TestStatus => {
     if (eventName === events.TEST_PASS) {
@@ -47,173 +46,11 @@ export const getStatus = (eventName: ValueOf<Testplane['events']>, events: Testp
     return TestStatus.IDLE;
 };
 
-const getSkipComment = (suite: TestplaneTestResult | TestplaneSuite): string | null | undefined => {
-    return suite.skipReason || suite.parent && getSkipComment(suite.parent);
-};
-
-const wrapSkipComment = (skipComment: string | null | undefined): string => {
-    return skipComment ?? 'Unknown reason';
-};
-
-const testItemsTheSame = (a: TestStepCompressed, b: TestStepCompressed): boolean => (
-    a &&
-    (!a[TestStepKey.Children] || !b[TestStepKey.Children]) &&
-    (!a[TestStepKey.IsFailed] || !b[TestStepKey.IsFailed]) &&
-    a[TestStepKey.Name] === b[TestStepKey.Name] &&
-    a[TestStepKey.Args].join() === b[TestStepKey.Args].join()
-);
-
-const arraysEqual = <I>(a: I[], b: I[], checkFunc: (a: I, b: I) => boolean): boolean => {
-    if (a.length !== b.length) {
-        return false;
-    }
-
-    return a.every((val, idx) => checkFunc(val, b[idx]));
-};
-
-const getTotalTime = (items: TestStepCompressed[], start: number, size: number): number => {
-    let total = 0;
-
-    for (let i = start; i < (start + size); i++) {
-        total += items[i][TestStepKey.Duration];
-    }
-
-    return total;
-};
-
-const getItemAverageTime = (
-    items: TestStepCompressed[],
-    start: number,
-    repeat: number,
-    index: number,
-    groupLen: number
-): number => {
-    let total = 0;
-
-    for (let i = 0; i < (repeat - 1); i++) {
-        total += items[start + (i * groupLen) + index][TestStepKey.Duration];
-    }
-
-    return parseFloat((total / (repeat - 1)).toFixed(2));
-};
-
-const MIN_REPEATS = 3; // Min count of repeats elements of group elements for squash
-
-const collapseRepeatingGroups = (
-    arr: TestStepCompressed[],
-    minRepeats: number = MIN_REPEATS
-): TestStepCompressed[] => {
-    const result: TestStepCompressed[] = [];
-    let i = 0;
-
-    while (i < arr.length) {
-        let foundGroup = false;
-
-        // max len of group can't be more that totalLen / minRepeats
-        for (let groupLen = 1; groupLen <= Math.floor((arr.length - i) / minRepeats); groupLen++) {
-            const group = arr.slice(i, i + groupLen);
-
-            let allGroupsMatch = true;
-
-            // check that group is repeated required count of times
-            for (let repeat = 1; repeat < minRepeats; repeat++) {
-                const nextGroupStart = i + repeat * groupLen;
-                const nextGroupEnd = nextGroupStart + groupLen;
-
-                if (nextGroupEnd > arr.length) {
-                    allGroupsMatch = false;
-                    break;
-                }
-
-                const nextGroup = arr.slice(nextGroupStart, nextGroupEnd);
-
-                if (!arraysEqual(group, nextGroup, testItemsTheSame)) {
-                    allGroupsMatch = false;
-                    break;
-                }
-            }
-
-            if (allGroupsMatch) {
-                foundGroup = true;
-                let repeatCount = minRepeats;
-
-                // finding another repeats of group
-                while (
-                    i + groupLen * repeatCount <= arr.length &&
-                    arraysEqual(
-                        group,
-                        arr.slice(i + groupLen * repeatCount, i + groupLen * (repeatCount + 1)),
-                        testItemsTheSame
-                    )
-                ) {
-                    repeatCount++;
-                }
-
-                const groupsTotalLen = groupLen * repeatCount;
-
-                if (groupLen === 1) {
-                    result.push({
-                        ...group[0],
-                        [TestStepKey.Duration]: getTotalTime(arr, i, groupsTotalLen),
-                        [TestStepKey.Repeat]: groupsTotalLen
-                    });
-                } else {
-                    result.push({
-                        [TestStepKey.Name]: 'Repeated group',
-                        [TestStepKey.Args]: [`${group.length} items`],
-                        [TestStepKey.Duration]: getTotalTime(arr, i, groupsTotalLen),
-                        [TestStepKey.TimeStart]: group[0][TestStepKey.TimeStart],
-                        [TestStepKey.IsFailed]: false,
-                        [TestStepKey.IsGroup]: true,
-                        [TestStepKey.Children]: group.map((item, index) => ({
-                            ...item,
-                            [TestStepKey.Repeat]: -1, // -1 need to detect in ui that this is child of group for show ~ in duration
-                            [TestStepKey.Duration]: getItemAverageTime(arr, i, repeatCount, index, groupLen)
-                        })),
-                        [TestStepKey.Repeat]: repeatCount
-                    });
-                }
-
-                i += groupsTotalLen;
-                break;
-            }
-        }
-
-        if (!foundGroup) {
-            result.push(arr[i]);
-            i++;
-        }
-    }
-
-    return result;
-};
-
-const getHistory = (history?: TestplaneTestResult['history']): TestStepCompressed[] => (
-    collapseRepeatingGroups(
-        history?.map((step) => {
-            const result: TestStepCompressed = {
-                [TestStepKey.Name]: step[TestStepKey.Name],
-                [TestStepKey.Args]: step[TestStepKey.Args],
-                [TestStepKey.Duration]: step[TestStepKey.Duration],
-                [TestStepKey.TimeStart]: step[TestStepKey.TimeStart],
-                [TestStepKey.IsFailed]: step[TestStepKey.IsFailed],
-                [TestStepKey.IsGroup]: step[TestStepKey.IsGroup]
-            };
-
-            if (step[TestStepKey.Children] && (step[TestStepKey.IsGroup] || step[TestStepKey.IsFailed])) {
-                result[TestStepKey.Children] = getHistory(step[TestStepKey.Children]);
-            }
-
-            return result;
-        }) ?? [],
-        MIN_REPEATS
-    )
-);
-
 export interface TestplaneTestResultAdapterOptions {
     attempt: number;
     status: TestStatus;
     duration: number;
+    saveHistoryMode?: Config['saveHistoryMode'];
 }
 
 export class TestplaneTestResultAdapter implements ReporterTestResult {
@@ -223,6 +60,7 @@ export class TestplaneTestResultAdapter implements ReporterTestResult {
     private _attempt: number;
     private _status: TestStatus;
     private _duration: number;
+    private _saveHistoryMode?: Config['saveHistoryMode'];
 
     static create(
         this: new (testResult: TestplaneTest | TestplaneTestResult, options: TestplaneTestResultAdapterOptions) => TestplaneTestResultAdapter,
@@ -232,7 +70,7 @@ export class TestplaneTestResultAdapter implements ReporterTestResult {
         return new this(testResult, options);
     }
 
-    constructor(testResult: TestplaneTest | TestplaneTestResult, {attempt, status, duration}: TestplaneTestResultAdapterOptions) {
+    constructor(testResult: TestplaneTest | TestplaneTestResult, {attempt, status, duration, saveHistoryMode}: TestplaneTestResultAdapterOptions) {
         this._testResult = testResult;
         this._errorDetails = null;
         this._timestamp = (this._testResult as TestplaneTestResult).startTime
@@ -246,6 +84,8 @@ export class TestplaneTestResultAdapter implements ReporterTestResult {
 
         this._attempt = attempt;
         this._duration = duration;
+
+        this._saveHistoryMode = saveHistoryMode;
     }
 
     get fullName(): string {
@@ -339,7 +179,7 @@ export class TestplaneTestResultAdapter implements ReporterTestResult {
     }
 
     get history(): TestStepCompressed[] {
-        return getHistory((this._testResult as TestplaneTestResult).history);
+        return getHistory((this._testResult as TestplaneTestResult).history, this._saveHistoryMode);
     }
 
     get error(): undefined | TestError {
